+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- [Add Fellow secrets helper commands](https://github.com/ManuelZierl/fellow/issues/83)
 - [add --version command](https://github.com/ManuelZierl/fellow/pull/92)
 
 ### Changed