+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [configure logging for memory and metadata](https://github.com/ManuelZierl/fellow/issues/32)
 - [restructure logging directory](https://github.com/ManuelZierl/fellow/issues/32)
 - [fixed typed badge by adding classifiers](https://github.com/ManuelZierl/fellow/issues/87)
+- [E2E Testing](https://github.com/ManuelZierl/fellow/issues/89)
 
 ### Changed
 - [use Path instead of str for paths in config](https://github.com/ManuelZierl/fellow/issues/32)