+++ CHANGELOG.md
@@ -8,6 +8,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+
+### Changed
+
+
+### Removed
+
+
+---
+
+## [0.1.0] - 2025-06-16
+
+### Added
 - [Add Fellow secrets helper commands](https://github.com/ManuelZierl/fellow/issues/83)
 - [add --version command](https://github.com/ManuelZierl/fellow/pull/92)
 - [introduce task_id for every run](https://github.com/ManuelZierl/fellow/issues/32)